Fix ul padding
Previously, I set the `ul` padding to be 0 in order to keep the blog
post list aligned the other text on the main page. However, this made
other `ul` elements appear to be pushed out of alignment to the
surrounding content.

This commit changes the behavior to have a padding on all `ul` elements
and creates a class for the list of posts so they are aligned to the
left edge.