Use Kramdown attributes for notes
I decided to use the kramdown gem's support for block attributes to make
my notes and blockquotes. It's much nicer than making a component for it.
The only thing is that I need to have the aria label in the block as well
to handle screen readers. But, that's fine for now.