Add social images to portfolio
I ran into an issue where I couldn't view the resource summary in the
default template. This meant that I needed a workaround and the only
thing I could think of was to create my own front matter excerpt of the
post and use that in the index pages.

I also added a banner_image_url to the front matter to create the image
meta tags.